+++ server/package.json
@@ -5,11 +5,16 @@
     "devDependencies": {
         "@types/cors": "^2.8.10",
         "@types/express": "^4.17.11",
+        "@types/jest": "^26.0.21",
         "@types/morgan": "^1.9.2",
         "@types/node": "^14.14.35",
+        "@types/supertest": "^2.0.10",
         "@types/swagger-ui-express": "^4.1.2",
+        "jest": "^26.6.3",
         "morgan": "^1.10.0",
         "nodemon": "^2.0.7",
+        "supertest": "^6.1.3",
+        "ts-jest": "^26.5.4",
         "ts-node": "^9.1.1",
         "typescript": "^4.2.3"
     },
@@ -24,6 +29,7 @@
     },
     "scripts": {
         "start": "ts-node src/index.ts",
-        "dev": "nodemon --exec ts-node src/index.ts"
+        "dev": "nodemon --exec ts-node src/index.ts",
+        "test": "jest"
     }
 }
